requirements.txt
line-bot-sdk==2.4.2
boto3==1.26.96
openai==0.27.2